Rename Core IConfigurationProvider to IConfigurationProviderService to avoid ambiguity
