+++ src/Integration.UnitTests/MefServices/ActiveSolutionBoundTrackerTests.cs
@@ -30,9 +30,9 @@ using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Integration.NewConnectedMode;
 using SonarQube.Client.Models;
 using SonarQube.Client;
-using IConfigurationProvider = SonarLint.VisualStudio.Integration.NewConnectedMode.IConfigurationProvider;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
