+++ src/Integration.UnitTests/ProfileConflicts/ConflictsManagerTests.cs
@@ -32,7 +32,7 @@ using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration.Binding;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
 using SonarLint.VisualStudio.Integration.ProfileConflicts;
-using IConfigurationProvider = SonarLint.VisualStudio.Integration.NewConnectedMode.IConfigurationProvider;
+
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
