+++ src/TestInfrastructure/Framework/ConfigurableConfigurationProvider.cs
@@ -22,7 +22,7 @@ using System;
 using FluentAssertions;
 using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
-using IConfigurationProvider = SonarLint.VisualStudio.Integration.NewConnectedMode.IConfigurationProvider;
+
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
